:sparkles: update playground-react form pages
